PYIC-7496: split out functional tests
